refactor: isolate connector registry type validation
